+++ R/plot_trait_distribution_beeswarm.R
@@ -2,7 +2,7 @@
 #' @description Plots distribution of trait values by a  grouping variable using ggbeeswarm package  
 #'
 #' @param austraits austraits data object
-#' @param plant_trait_name Name of trait to plot
+#' @param trait_name Name of trait to plot
 #' @param y_axis_category One of `dataset_id`, `family`
 #' @param highlight specify a group to highlight
 #' @param hide_ids add label on y_axis?
@@ -17,23 +17,23 @@
 #' @export
 
 #
-plot_trait_distribution_beeswarm <- function(austraits, 
-                                             plant_trait_name, 
-                                             y_axis_category, 
-                                             highlight=NA, 
+plot_trait_distribution_beeswarm <- function(austraits,
+                                             trait_name,
+                                             y_axis_category,
+                                             highlight=NA,
                                              hide_ids = FALSE) {
   
   # Check compatability
   status <- check_compatibility(austraits)
   
   # If compatible
-  if(!status){
+  if(!status) {
     function_not_supported(austraits)
-  } 
+  }
   # Subset data to this trait
-  austraits_trait <- extract_trait(austraits, plant_trait_name)
+  austraits_trait <- extract_trait(austraits, trait_name)
   
-  my_shapes = c("_min" = 60, "_mean" = 16, "_max" =62, "unknown" = 18)
+  my_shapes <- c("_min" = 60, "_mean" = 16, "_max" = 62, "unknown" = 18)
   
   as_shape <- function(value_type) {
     p <- rep("unknown", length(value_type))
@@ -46,19 +46,21 @@ plot_trait_distribution_beeswarm <- function(austraits,
   
   tax_info  <- austraits_trait$taxa %>% dplyr::select(taxon_name, family)
   
-  data <- 
+  data <-
     austraits_trait$traits %>%
     dplyr::mutate(shapes = as_shape(value_type)) %>%
-    dplyr::left_join(by = "taxon_name", tax_info)
+    dplyr::left_join(by = "taxon_name", tax_info) %>%
+    dplyr::mutate(value = as.numeric(value))
   
   # Define grouping variables and derivatives
-  if(!y_axis_category %in% names(data)){
+  if(!y_axis_category %in% names(data)) {
     stop("Incorrect grouping variable! Currently implemented for `family` or `dataset_id`")
   }
   
   # define grouping variable, ordered by group-level by mean values
   # use log_value where possible
-  if(min(data$value, na.rm=TRUE) > 0 ) {
+  
+  if(min(data$value, na.rm=TRUE) > 0) {
     data$value2 <- log10(data$value)
   } else {
     data$value2 <- data$value
@@ -77,8 +79,8 @@ plot_trait_distribution_beeswarm <- function(austraits,
   }
   
 
-  vals <- list(minimum = purrr::pluck(austraits_trait, "definitions", plant_trait_name, "allowed_values_min"),
-           maximum = purrr::pluck(austraits_trait, "definitions", plant_trait_name, "allowed_values_max"))
+  vals <- list(minimum = purrr::pluck(austraits_trait, "definitions", trait_name, "allowed_values_min"),
+           maximum = purrr::pluck(austraits_trait, "definitions", trait_name, "allowed_values_max"))
   
   range <- (vals$maximum/vals$minimum)
   
@@ -143,14 +145,14 @@ plot_trait_distribution_beeswarm <- function(austraits,
                     labels = scientific_10,
                     limits=c(vals$minimum, vals$maximum))
     p2 <- p2 +
-      ggplot2::scale_x_log10(name=paste(plant_trait_name, ' (', data$unit[1], ')'),
+      ggplot2::scale_x_log10(name=paste(trait_name, ' (', data$unit[1], ')'),
                     breaks = scales::breaks_log(),
                     labels = scientific_10,
                     limits=c(vals$minimum, vals$maximum))
   } else {
     p1 <- p1 + ggplot2::scale_x_continuous(limits=c(vals$minimum, vals$maximum))
     p2 <- p2 + ggplot2::scale_x_continuous(limits=c(vals$minimum, vals$maximum)) +
-      ggplot2::xlab(paste(plant_trait_name, ' (', data$unit[1], ')'))
+      ggplot2::xlab(paste(trait_name, ' (', data$unit[1], ')'))
     
   }