+++ tests/testthat/test-extract_.R
@@ -1,9 +1,12 @@
 not_supported_austraits <- list(austraits_3.0.2_lite, austraits_4.2.0_lite) 
 
 dataset_id = "Falster_2003"
+dataset_id2 = "Cernusak_2006"
+dataset_id3 = "Wright_2019"
 trait_name = "leaf_area"
 family = "Rubiaceae"
 genus = "Eucalyptus"
+taxon_name = "Banskia serrata"
 
 test_that("Error message is triggered", {
   expect_error(austraits_5.0.0_lite |> extract_taxa())
@@ -28,17 +31,26 @@ test_that("Function runs", {
 }
 )
 
+test_that("Function runs", {
+  expect_visible(austraits_5.0.0_lite %>% extract_taxa(family = family))
+  expect_visible(austraits_5.0.0_lite %>% extract_taxa(genus = genus))
+  expect_visible(extract_dataset(austraits_5.0.0_lite, dataset_id = dataset_id))
+  expect_visible(extract_trait(austraits_5.0.0_lite, trait_names = trait_name))
+}
+)
+
+
 test_that("extracted dataset has some structure as austraits build", {
   subset <- extract_dataset(austraits_5.0.0_lite, dataset_id = dataset_id)
   trait_subset <- extract_trait(austraits_5.0.0_lite, trait_names = trait_name)
   
   expect_s3_class(austraits_5.0.0_lite, "austraits")
-  expect_equal(length(austraits_5.0.0_lite), length(subset))
-  expect_equal(sort(names(austraits_5.0.0_lite)), sort(names(subset)))
+  expect_equal(length(subset), length(austraits_5.0.0_lite))
+  expect_equal(sort(names(subset)), sort(names(austraits_5.0.0_lite)))
   
-  expect_equal(length(austraits_5.0.0_lite), length(trait_subset))
-  expect_equal(sort(names(austraits_5.0.0_lite)), sort(names(trait_subset)))
-  expect_named(austraits_5.0.0_lite, names(trait_subset))  
+  expect_equal(length(trait_subset), length(austraits_5.0.0_lite))
+  expect_equal(sort(names(trait_subset)), sort(names(austraits_5.0.0_lite)))
+  expect_equal(names(trait_subset), names(austraits_5.0.0_lite))  
   
   expect_type(austraits_5.0.0_lite %>% extract_taxa(family = family), "list")
   expect_type(austraits_5.0.0_lite %>% extract_taxa(genus = genus), "list")           
@@ -52,20 +64,137 @@ test_that("extracted dataset has some structure as austraits build", {
   expect_equal(test_fam$taxa$family %>% unique(), family)
 })
 
+test_that("extracts using generalised extract function behaves as expected - extracting by dataset_id", {
+  
+  subset_by_dataset_id <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "dataset_id", col_value = dataset_id)
+  expect_equal(length(austraits_5.0.0_lite), length(subset_by_dataset_id))
+  expect_equal(nrow(subset_by_dataset_id$locations), nrow(austraits_5.0.0_lite$locations %>% dplyr::filter(dataset_id == "Falster_2003")))
+  expect_equal(nrow(subset_by_dataset_id$contexts), nrow(austraits_5.0.0_lite$contexts %>% dplyr::filter(dataset_id == "Falster_2003")))
+  
+  subset_by_dataset_id2 <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "dataset_id", col_value = dataset_id2)
+  expect_equal(length(austraits_5.0.0_lite), length(subset_by_dataset_id2))
+  expect_equal(nrow(subset_by_dataset_id2$locations), nrow(austraits_5.0.0_lite$locations %>% dplyr::filter(dataset_id == "Cernusak_2006")))
+  expect_equal(nrow(subset_by_dataset_id2$contexts), nrow(austraits_5.0.0_lite$contexts %>% dplyr::filter(dataset_id == "Cernusak_2006")))
+  expect_equal(nrow(subset_by_dataset_id2$methods), nrow(austraits_5.0.0_lite$methods %>% dplyr::filter(dataset_id == "Cernusak_2006")))
+  expect_equal(nrow(subset_by_dataset_id2$contributors), nrow(austraits_5.0.0_lite$contributors %>% dplyr::filter(dataset_id == "Cernusak_2006")))
+  expect_equal(names(subset_by_dataset_id2), names(austraits_5.0.0_lite))
+  
+  subset_by_dataset_id3 <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "dataset_id", col_value = dataset_id3)
+  expect_equal(length(austraits_5.0.0_lite), length(subset_by_dataset_id3))
+  expect_equal(nrow(subset_by_dataset_id3$locations), nrow(austraits_5.0.0_lite$locations %>% dplyr::filter(dataset_id == "Wright_2019")))
+  expect_equal(nrow(subset_by_dataset_id3$contexts), nrow(austraits_5.0.0_lite$contexts %>% dplyr::filter(dataset_id == "Wright_2019")))
+  expect_equal(nrow(subset_by_dataset_id3$methods), nrow(austraits_5.0.0_lite$methods %>% dplyr::filter(dataset_id == "Wright_2019")))
+  expect_equal(nrow(subset_by_dataset_id3$contributors), nrow(austraits_5.0.0_lite$contributors %>% dplyr::filter(dataset_id == "Wright_2019")))
+  expect_equal(names(subset_by_dataset_id3), names(austraits_5.0.0_lite))
+  })
+
+
+test_that("that you can link two calls of `extract_data` together", {
+  subset_by_dataset_id2 <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "dataset_id", col_value = dataset_id2)
+  expect_no_error(extract_data(database = subset_by_dataset_id2, table = "traits", col = "trait_name", col_value = "leaf_mass_per_area"))
+  
+  subset_by_dataset_id3 <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "trait_name", col_value = "leaf_mass_per_area")
+  expect_no_error(extract_data(database = subset_by_dataset_id3, table = "contexts", col = "context_property", col_value = "age"))
+  subset_by_dataset_id_and_context <- extract_data(database = subset_by_dataset_id3, table = "contexts", col = "context_property", col_value = "age")
+  expect_gt(nrow(subset_by_dataset_id3$contexts), nrow(subset_by_dataset_id_and_context$contexts))
+  })
+  
+test_that("extracts using generalised extract function behaves as expected - extracting by `life_stage", {
+  
+  life_stage_test <- "sapling"
+  
+  subset_by_age_class <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "life_stage", col_value = life_stage_test)
+  datasets_in_subset <- subset_by_age_class$traits %>% dplyr::distinct(dataset_id)
+  
+  expect_lt(nrow(subset_by_age_class$locations), nrow(austraits_5.0.0_lite$locations))
+  expect_lt(nrow(subset_by_age_class$contexts), nrow(austraits_5.0.0_lite$contexts))
+  expect_contains(datasets_in_subset$dataset_id, subset_by_age_class$contexts$dataset_id)
+  expect_contains(datasets_in_subset$dataset_id, subset_by_age_class$locations$dataset_id)
+  expect_equal(names(subset_by_age_class), names(austraits_5.0.0_lite))
+}) 
+
+test_that("extracts using generalised extract function behaves as expected - extracting by `location_property", {
+  
+  location_property_test <- "temperature"
+  
+  subset_by_location_property <- extract_data(database = austraits_5.0.0_lite, table = "locations", col = "location_property", col_value = location_property_test)
+  datasets_in_subset <- subset_by_location_property$locations %>% dplyr::distinct(dataset_id)
+  
+  expect_lt(nrow(subset_by_location_property$traits), nrow(austraits_5.0.0_lite$traits))
+  expect_lt(nrow(subset_by_location_property$locations), nrow(austraits_5.0.0_lite$locations))
+  expect_lt(nrow(subset_by_location_property$contexts), nrow(austraits_5.0.0_lite$contexts))
+  expect_contains(datasets_in_subset$dataset_id, subset_by_location_property$locations$dataset_id)
+  # for locations, all datasets in subset, since it is the starting point
+  expect_contains(subset_by_location_property$locations$dataset_id, datasets_in_subset$dataset_id)
+  expect_contains(datasets_in_subset$dataset_id, subset_by_location_property$traits$dataset_id)
+  # similarly for traits, all datasets from locations must be in traits
+  expect_contains(subset_by_location_property$traits$dataset_id, datasets_in_subset$dataset_id)
+  # however contexts will only be a subset of dataset_ids, so only 1 direction is true
+  expect_contains(datasets_in_subset$dataset_id, subset_by_location_property$contexts$dataset_id)
+  expect_equal(names(subset_by_location_property), names(austraits_5.0.0_lite))
+})  
+
+
+test_that("extracts for which there are no matches work`", {
+  context_property_test <- "platypus"
+  expect_message(extract_data(database = austraits_5.0.0_lite,  table = "contexts", col = "context_property", col_value = context_property_test))
+  expect_equal(nrow(extract_data(database = austraits_5.0.0_lite,  table = "contexts", col = "context_property", col_value = context_property_test)$traits), 0)
+  
+  location_property_test <- "green flowers"
+  expect_message(extract_data(database = austraits_5.0.0_lite,  table = "locations", col = "location_property", col_value = location_property_test))
+  expect_equal(nrow(extract_data(database = austraits_5.0.0_lite,  table = "locations", col = "location_property", col_value = location_property_test)$traits), 0)
+})
+  
+test_that("extracts using generalised extract function behaves as expected - extracting by `context_property`", {
+  
+  context_property_test <- "fire"
+  
+  subset_by_context_property <- extract_data(database = austraits_5.0.0_lite,  table = "contexts", col = "context_property", col_value = context_property_test)
+
+  datasets_in_subset <- subset_by_context_property$contexts %>% dplyr::distinct(dataset_id)
+  subset_using_filter <- austraits_5.0.0_lite$contexts %>% dplyr::filter(stringr::str_detect(context_property, context_property_test))
+  
+  expect_lt(nrow(subset_by_context_property$traits), nrow(austraits_5.0.0_lite$traits))
+  expect_lt(nrow(subset_by_context_property$locations), nrow(austraits_5.0.0_lite$locations))
+  expect_lt(nrow(subset_by_context_property$contexts), nrow(austraits_5.0.0_lite$contexts))
+  # for locations, all datasets in subset, since it is the starting point
+  expect_contains(datasets_in_subset$dataset_id, subset_by_context_property$contexts$dataset_id)
+  expect_contains(subset_by_context_property$contexts$dataset_id, datasets_in_subset$dataset_id)
+  # similarly for traits, all datasets from locations must be in traits
+  expect_contains(datasets_in_subset$dataset_id, subset_by_context_property$traits$dataset_id)
+  expect_contains(subset_by_context_property$traits$dataset_id, datasets_in_subset$dataset_id)
+  # however contexts will only be a subset of dataset_ids, so only 1 direction is true
+  expect_contains(datasets_in_subset$dataset_id, subset_by_context_property$locations$dataset_id)
+  # can't work out a way to make this run, but for locations, not all datasets will be represented
+  # expect_error(subset_by_context_property$locations$dataset_id, datasets_in_subset$dataset_id)
+  expect_equal(names(subset_by_context_property), names(austraits_5.0.0_lite))
+  
+  # this should be true, because the proper extract function also retains other context property data linked to the same observation
+  expect_gte(nrow(subset_by_context_property$contexts), nrow(subset_using_filter))
+  # however both methods should be including the same dataset_id's 
+  expect_equal(
+    subset_using_filter %>% dplyr::distinct(dataset_id) %>% dplyr::arrange(dataset_id), 
+    subset_by_context_property$contexts %>% dplyr::distinct(dataset_id) %>% dplyr::arrange(dataset_id)
+    )
+})  
 
 test_that("Extraction of dataset was successful", {
   subset <- extract_dataset(austraits_5.0.0_lite, dataset_id = dataset_id)
   trait_subset <- extract_trait(austraits_5.0.0_lite, trait_names = trait_name)
-  
   expect_match(dataset_id, unique(subset$traits$dataset_id))
   expect_equal(1, dplyr::n_distinct(subset$traits$dataset_id))
+  # this isn't an exact match, because the matches use partial string matches and for this example, both `leaf_area` and `leaf_area_ratio` being matched
   expect_match(trait_name, unique(trait_subset$traits$trait_name))
-  expect_equal(1, dplyr::n_distinct(trait_subset$traits$trait_name))  
-  
-  expect_match(dataset_id, unique(subset$traits$dataset_id))
-  expect_equal(1, dplyr::n_distinct(subset$traits$dataset_id))
-  expect_match(trait_name, unique(trait_subset$traits$trait_name))
-  expect_equal(1, dplyr::n_distinct(trait_subset$traits$trait_name)) 
+  expect_equal(1, dplyr::n_distinct(trait_subset$traits$trait_name))
 })
 
+test_that("Extraction of dataset was successful using `extract_data`", {
+  subset <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "dataset_id", col_value = dataset_id)
+  trait_subset <- extract_data(database = austraits_5.0.0_lite, table = "traits", col = "trait_name", col_value = trait_name)
+  expect_match(dataset_id, unique(subset$traits$dataset_id))
+  expect_equal(dplyr::n_distinct(subset$traits$dataset_id), 1)
+  expect_contains(unique(trait_subset$traits$trait_name), trait_name)
+  expect_equal(dplyr::n_distinct(trait_subset$traits$trait_name), 1)
+  expect_equal(trait_subset$traits %>% dplyr::distinct(dataset_id) %>% nrow(), 8) #something weird here, keeps isolating between 8 & 5
+})